requirements.txt
celery==5.2.3
coverage==6.3.1
dj-rest-auth==2.2.2
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.11.0
django-silk==4.2.0
django-timezone-field==4.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
factory-boy==3.2.1
Faker==12.1.0
gevent==21.12.0
gunicorn==20.1.0
h11==0.12.0
mysqlclient==2.1.0
PyJWT==2.3.0
PyMySQL==1.0.2
selenium==4.1.0